requirements.txt
astroid==2.4.2
attrs==20.2.0
click==7.1.2
et-xmlfile==1.0.1
Faker==4.1.2
Flask==1.1.2
importlib-metadata==1.7.0
iniconfig==1.0.1
isort==5.5.3
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
numpy==1.19.2
openpyxl==3.0.5
packaging==20.4
pandas==1.1.2
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
XlsxWriter==1.3.4
zipp==3.1.0
astroid==2.4.2
attrs==20.2.0
click==7.1.2
et-xmlfile==1.0.1
Faker==4.1.2
Flask==1.1.2
importlib-metadata==1.7.0
iniconfig==1.0.1
isort==5.5.3
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
numpy==1.19.2
openpyxl==3.0.5
packaging==20.4
pandas==1.1.2
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.4
zipp==3.1.0
